自带配置脚本适配init动作

+++ app/src/main/assets/powercfg/msm8953/powercfg-default.sh
@@ -1,6 +1,10 @@
 #!/system/bin/sh
 
 action=$1
+if [[ "$action" = "init" ]] && [[ -f '/data/powercfg-base.sh' ]]; then
+    sh /data/powercfg-base.sh
+	exit 0
+fi
 
 if [ ! `cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_governor` = "interactive" ]; then
 	echo 'interactive' > /sys/devices/system/cpu/cpu0/cpufreq/scaling_governor
